requirements.txt
# Core LLM / RAG
langchain==0.3.7
langchain-openai==0.2.9
langchain-community==0.3.7
langchain-chroma==0.1.4
openai==1.54.4

# Vector Store
chromadb==0.5.18

# Graph Database
neo4j==5.26.0          # Memgraph uses the Neo4j Bolt protocol

# API
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Data / Utils
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
tenacity==9.0.0
tiktoken==0.8.0

# Evaluation
ragas==0.2.6
datasets==3.1.0
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2

# Dev / Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
